file structure : moved gameMechanics file to utilities folder
