+++ src/components/Game.jsx
@@ -5,7 +5,7 @@ import ActionMenu from "./ActionMenu"
 import DialogBox from "./DialogBox"
 import Notepad from "./Notepad"
 import useStore from "./store"
-import { InitializeGame } from "./gameMechanics"
+import { InitializeGame } from "../utilities/gameMechanics"
 import DevConsole from "./DevConsole"
 import Timer from "./Timer"
 import PlayerChangeRoom from "./RoomChanger"
